Allow multiple agencies, in a single gtfs, as well as multiple gtfs files.
Instead of storing agency_id in every table, we store the feed_name. The feed
name is derived from the zip filename or the directory name supplied to `load`.
This allows us to store multiple GTFS files in one database, as long as their
names differ, but it also allows to have more than one agency per gtfs.