removed the unicode casting
The unicode casting was really slowing us down. So it was removed,
under the assumption that all given 'strings' will be unicode (as in python3).
Also, in feed.py on some cases encoding-sig was called before definition!
so it was also removed. The GTFS spec say everything would be unicode anyway.